rename from tf to jet - remaining small parts
