+++ internal/controller/providerconfig/config.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package awsjet
+package providerconfig
 
 import (
 	"github.com/crossplane-contrib/terrajet/pkg/config"
