+++ package/crds/aws.jet.crossplane.io_providerconfigs.yaml
@@ -29,7 +29,7 @@ spec:
     name: v1alpha1
     schema:
       openAPIV3Schema:
-        description: A ProviderConfig configures a TF AWS provider.
+        description: A ProviderConfig configures a Jet AWS provider.
         properties:
           apiVersion:
             description: 'APIVersion defines the versioned schema of this representation
